functional component and restructuring
