+++ src/styled-components/Animations.styled.js
