+++ src/styled-components/Board.styled.js
